requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright>=1.48.0
beautifulsoup4==4.12.3
lxml>=4.9.0
openai==1.10.0
httpx==0.26.0
Pillow>=10.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
